pyproject.toml
[project]
name = "finagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai==1.12.1",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv==1.1.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
